# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dc0d95e9ef7a1181f8e1322
  title: khachsan
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /user/:
    post:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1334
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employess:
                  type: integer
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1330
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1336
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1332
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e132f
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  bill  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
    User:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-id-db: 5dc0d95e9ef7a1181f8e1323
      x-skaffolder-id-entity: 5dc0d95e9ef7a1181f8e1324
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        employess:
          type: integer
          x-skaffolder-id-attr: 5dc0eb1d9ef7a1181f8e1413
          x-skaffolder-type: Integer
        mail:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1327
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1328
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1326
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e132a
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1329
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1325
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...